+++ typed_floats/Cargo.toml
@@ -16,7 +16,7 @@ rust-version.workspace = true
 bench = false
 
 [features]
-default = ["std", "compiler_hints"]
+default = ["std", "compiler_hints", "f32", "f64"]
 # follows each debug_assert! with core::hint::unreachable_unchecked
 compiler_hints = []
 # Add the `serde` dependency to enable serialization and deserialization of the types.
@@ -29,6 +29,10 @@ libm = ["dep:num-traits", "typed_floats_macros/libm"]
 # This will override the `compiler_hints` feature, and adds a little overhead to `new_unchecked`.
 # This feature can be enabled by any parent crate to ensure no undefined behavior.
 ensure_no_undefined_behavior = []
+# Support for 32-bit floats.
+f32 = ["typed_floats_macros/f32"]
+# Support for 64-bit floats.
+f64 = ["typed_floats_macros/f64"]
 
 [dependencies]
 rustversion.workspace = true
