+++ typed_floats_macros/src/lib.rs
@@ -216,13 +216,19 @@ pub fn generate_docs(input: proc_macro::TokenStream) -> proc_macro::TokenStream 
 /// Generate the `PartialEq`, `From` and `TryFrom` implementations.
 #[proc_macro]
 pub fn generate_floats(_input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    let floats_f64 = get_definitions("f64");
-    let floats_f32 = get_definitions("f32");
-
     let mut output = proc_macro2::TokenStream::new();
 
-    output.extend(do_generate_floats(&floats_f64));
-    output.extend(do_generate_floats(&floats_f32));
+    #[cfg(feature = "f64")]
+    {
+        let floats_f64 = get_definitions("f64");
+        output.extend(do_generate_floats(&floats_f64));
+    }
+
+    #[cfg(feature = "f32")]
+    {
+        let floats_f32 = get_definitions("f32");
+        output.extend(do_generate_floats(&floats_f32));
+    }
 
     output.into()
 }
